Don't squeeze list of cellids in cache.load
This fixes a bug that occurs when there is only one cell.
This squeeze is no longer necessary because we have done squeeze
before saving since a6194c5. Only caches created before then are
problematic---and if you have to use a cache that old, you should
probably use an old revision of CellProfiler-Analyst as well.